+++ Packages/DocBlockr/jsdocs.py
@@ -1,8 +1,9 @@
 """
-DocBlockr v2.7.4
+DocBlockr v2.9.1
 by Nick Fisher
 https://github.com/spadgos/sublime-jsdocs
 """
+import sublime
 import sublime_plugin
 import re
 import string
@@ -32,7 +33,7 @@ def counter():
 
 
 def escape(str):
-    return string.replace(str, '$', '\$')
+    return str.replace('$', '\$').replace('{', '\{').replace('}', '\}')
 
 
 def is_numeric(val):
@@ -45,15 +46,20 @@ def is_numeric(val):
 
 def getParser(view):
     scope = view.scope_name(view.sel()[0].end())
+    res = re.search('\\bsource\\.([a-z+]+)', scope)
+    sourceLang = res.group(1) if res else 'js'
     viewSettings = view.settings()
 
-    if re.search("source\\.php", scope):
+    if sourceLang == "php":
         return JsdocsPHP(viewSettings)
-    elif re.search("source\\.coffee", scope):
+    elif sourceLang == "coffee":
         return JsdocsCoffee(viewSettings)
-    elif re.search("source\\.actionscript", scope):
+    elif sourceLang == "actionscript":
         return JsdocsActionscript(viewSettings)
-
+    elif sourceLang == "c++" or sourceLang == 'c':
+        return JsdocsCPP(viewSettings)
+    elif sourceLang == 'objc' or sourceLang == 'objc++':
+        return JsdocsObjC(viewSettings)
     return JsdocsJavascript(viewSettings)
 
 
@@ -61,9 +67,16 @@ class JsdocsCommand(sublime_plugin.TextCommand):
 
     def run(self, edit, inline=False):
         v = self.view
-
         settings = v.settings()
         point = v.sel()[0].end()
+
+        # trailing characters are put inside the body of the comment
+        trailingRgn = sublime.Region(point, v.line(point).end())
+        trailingString = v.substr(trailingRgn).strip()
+        # drop trailing '*/'
+        trailingString = escape(re.sub('\\s*\\*\\/\\s*$', '', trailingString))
+        # erase characters in the view (will be added to the output later)
+        v.erase(edit, trailingRgn)
 
         indentSpaces = " " * max(0, settings.get("jsdocs_indentation_spaces", 1))
         prefix = "*"
@@ -75,14 +88,18 @@ class JsdocsCommand(sublime_plugin.TextCommand):
         parser = getParser(v)
         parser.inline = inline
 
+        # use trailing string as a description of the function
+        if trailingString:
+            parser.setNameOverride(trailingString)
+
         # read the next line
-        line = read_line(v, point + 1)
+        line = parser.getDefinition(v, point + 1)
         out = None
 
         # if there is a line following this
         if line:
             if parser.isExistingComment(line):
-                write(v, "\n *" + (" " * indentSpaces))
+                write(v, "\n *" + indentSpaces)
                 return
             # match against a function declaration.
             out = parser.parse(line)
@@ -161,23 +178,31 @@ class JsdocsCommand(sublime_plugin.TextCommand):
                 for line in out:
                     snippet += "\n " + prefix + (indentSpaces + line if line else "")
             else:
-                snippet += "\n " + prefix + indentSpaces + "$0"
+                snippet += "\n " + prefix + indentSpaces + "${0:" + trailingString + '}'
 
             snippet += "\n" + closer
             write(v, snippet)
 
 
-class JsdocsParser:
+class JsdocsParser(object):
 
     def __init__(self, viewSettings):
         self.viewSettings = viewSettings
         self.setupSettings()
+        self.nameOverride = None
 
     def isExistingComment(self, line):
         return re.search('^\\s*\\*', line)
 
+    def setNameOverride(self, name):
+        """ overrides the description of the function - used instead of parsed description """
+        self.nameOverride = name
+
+    def getNameOverride(self):
+        return self.nameOverride
+
     def parse(self, line):
-        out = self.parseFunction(line)  # (name, args, options)
+        out = self.parseFunction(line)  # (name, args, retval, options)
         if (out):
             return self.formatFunction(*out)
 
@@ -212,13 +237,14 @@ class JsdocsParser:
 
         return out
 
-    def formatFunction(self, name, args, options={}):
+    def formatFunction(self, name, args, retval, options={}):
         out = []
         if 'as_setter' in options:
             out.append('@private')
             return out
 
-        out.append("${1:[%s description]}" % (name))
+        description = self.getNameOverride() or ('[%s description]' % escape(name))
+        out.append("${1:%s}" % description)
 
         self.addExtraTags(out)
 
@@ -226,20 +252,22 @@ class JsdocsParser:
         if (args):
             # remove comments inside the argument list.
             args = re.sub("/\*.*?\*/", '', args)
-            for arg in self.parseArgs(args):
+            for argType, argName in self.parseArgs(args):
                 typeInfo = ''
                 if self.settings['typeInfo']:
                     typeInfo = '%s${1:%s}%s ' % (
                         "{" if self.settings['curlyTypes'] else "",
-                        escape(arg[0] or self.guessTypeFromName(arg[1]) or "[type]"),
+                        escape(argType or self.guessTypeFromName(argName) or "[type]"),
                         "}" if self.settings['curlyTypes'] else "",
                     )
                 out.append("@param %s%s ${1:[description]}" % (
                     typeInfo,
-                    escape(arg[1])
+                    escape(argName)
                 ))
 
-        retType = self.getFunctionReturnType(name)
+        # return value type might be already available in some languages but
+        # even then ask language specific parser if it wants it listed
+        retType = self.getFunctionReturnType(name, retval)
         if retType is not None:
             typeInfo = ''
             if self.settings['typeInfo']:
@@ -255,32 +283,40 @@ class JsdocsParser:
 
             if (self.viewSettings.get('jsdocs_return_description')):
                 format_str = "%s%s %s${1:[description]}"
+                third_arg = ""
 
                 # the extra space here is so that the description will align with the param description
-                format_args.append(" " if (args and not self.viewSettings.get('jsdocs_per_section_indent')) else "")
+                if args and self.viewSettings.get('jsdocs_align_tags') == 'deep':
+                    if not self.viewSettings.get('jsdocs_per_section_indent'):
+                        third_arg = " "
+
+                format_args.append(third_arg)
             else:
                 format_str = "%s%s"
 
             out.append(format_str % tuple(format_args))
 
+        for notation in self.getMatchingNotations(name):
+            if 'tags' in notation:
+                out.extend(notation['tags'])
+
         return out
 
-    def getFunctionReturnType(self, name):
+    def getFunctionReturnType(self, name, retval):
         """ returns None for no return type. False meaning unknown, or a string """
-        name = re.sub("^[$_]", "", name)
 
         if re.match("[A-Z]", name):
             # no return, but should add a class
             return None
 
-        if re.match('(?:set|add)($|[A-Z_])', name):
+        if re.match('[$_]?(?:set|add)($|[A-Z_])', name):
             # setter/mutator, no return
             return None
 
-        if re.match('(?:is|has)($|[A-Z_])', name):  # functions starting with 'is' or 'has'
+        if re.match('[$_]?(?:is|has)($|[A-Z_])', name):  # functions starting with 'is' or 'has'
             return self.settings['bool']
 
-        return False
+        return self.guessTypeFromName(name) or False
 
     def parseArgs(self, args):
         """ an array of tuples, the first being the best guess at the type, the second being the name """
@@ -302,18 +338,11 @@ class JsdocsParser:
             out.extend(extraTags)
 
     def guessTypeFromName(self, name):
-        hungarian_map = self.viewSettings.get('jsdocs_notation_map', [])
-        if len(hungarian_map):
-            for rule in hungarian_map:
-                matched = False
-                if 'prefix' in rule:
-                    matched = re.match(rule['prefix'] + "[A-Z_]", name)
-                elif 'regex' in rule:
-                    matched = re.search(rule['regex'], name)
-
-                if matched:
-
-                    return self.settings[rule['type']] if rule['type'] in self.settings else rule['type']
+        matches = self.getMatchingNotations(name)
+        if len(matches):
+            rule = matches[0]
+            if ('type' in rule):
+                return self.settings[rule['type']] if rule['type'] in self.settings else rule['type']
 
         if (re.match("(?:is|has)[A-Z_]", name)):
             return self.settings['bool']
@@ -323,18 +352,63 @@ class JsdocsParser:
 
         return False
 
+    def getMatchingNotations(self, name):
+        def checkMatch(rule):
+            if 'prefix' in rule:
+                regex = re.escape(rule['prefix'])
+                if re.match('.*[a-z]', rule['prefix']):
+                    regex += '(?:[A-Z_]|$)'
+                return re.match(regex, name)
+            elif 'regex' in rule:
+                return re.search(rule['regex'], name)
+
+        return filter(checkMatch, self.viewSettings.get('jsdocs_notation_map', []))
+
+    def getDefinition(self, view, pos):
+        """
+        get a relevant definition starting at the given point
+        returns string
+        """
+        maxLines = 25  # don't go further than this
+        openBrackets = 0
+
+        definition = ''
+
+        def countBrackets(total, bracket):
+            return total + (1 if bracket == '(' else -1)
+
+        for i in xrange(0, maxLines):
+            line = read_line(view, pos)
+            if line is None:
+                break
+
+            pos += len(line) + 1
+            # strip comments
+            line = re.sub("//.*", "", line)
+            line = re.sub(r"/\*.*\*/", "", line)
+            if definition == '':
+                if not self.settings['fnOpener'] or not re.search(self.settings['fnOpener'], line):
+                    definition = line
+                    break
+            definition += line
+            openBrackets = reduce(countBrackets, re.findall('[()]', line), openBrackets)
+            if openBrackets == 0:
+                break
+        return definition
+
 
 class JsdocsJavascript(JsdocsParser):
     def setupSettings(self):
+        identifier = '[a-zA-Z_$][a-zA-Z_$0-9]*'
         self.settings = {
             # curly brackets around the type information
             "curlyTypes": True,
             'typeInfo': True,
             "typeTag": "type",
             # technically, they can contain all sorts of unicode, but w/e
-            "varIdentifier": '[a-zA-Z_$][a-zA-Z_$0-9]*',
-            "fnIdentifier": '[a-zA-Z_$][a-zA-Z_$0-9]*',
-
+            "varIdentifier": identifier,
+            "fnIdentifier":  identifier,
+            "fnOpener": 'function(?:\\s+' + identifier + ')?\\s*\\(',
             "commentCloser": " */",
             "bool": "Boolean",
             "function": "Function"
@@ -348,17 +422,17 @@ class JsdocsJavascript(JsdocsParser):
             # function fnName
             + '(?:\s+(?P<name2>' + self.settings['fnIdentifier'] + '))?'
             # (arg1, arg2)
-            + '\s*\((?P<args>.*)\)',
+            + '\s*\(\s*(?P<args>.*)\)',
             line
         )
         if not res:
             return None
 
         # grab the name out of "name1 = function name2(foo)" preferring name1
-        name = escape(res.group('name1') or res.group('name2') or '')
+        name = res.group('name1') or res.group('name2') or ''
         args = res.group('args')
 
-        return (name, args)
+        return (name, args, None)
 
     def parseVar(self, line):
         res = re.search(
@@ -406,6 +480,7 @@ class JsdocsPHP(JsdocsParser):
             'typeTag': "var",
             'varIdentifier': '[$]' + nameToken + '(?:->' + nameToken + ')*',
             'fnIdentifier': nameToken,
+            'fnOpener': 'function(?:\\s+' + nameToken + ')?\\s*\\(',
             'commentCloser': ' */',
             'bool': "boolean",
             'function': "function"
@@ -417,13 +492,13 @@ class JsdocsPHP(JsdocsParser):
             + '(?P<name>' + self.settings['fnIdentifier'] + ')'
             # function fnName
             # (arg1, arg2)
-            + '\\s*\\((?P<args>.*)\)',
+            + '\\s*\\(\\s*(?P<args>.*)\)',
             line
         )
         if not res:
             return None
 
-        return (res.group('name'), res.group('args'))
+        return (res.group('name'), res.group('args'), None)
 
     def getArgType(self, arg):
         #  function add($x, $y = 1)
@@ -481,7 +556,7 @@ class JsdocsPHP(JsdocsParser):
             return res and res.group(1) or None
         return None
 
-    def getFunctionReturnType(self, name):
+    def getFunctionReturnType(self, name, retval):
         if (name[:2] == '__'):
             if name in ('__construct', '__destruct', '__set', '__unset', '__wakeup'):
                 return None
@@ -491,19 +566,72 @@ class JsdocsPHP(JsdocsParser):
                 return 'string'
             if name == '__isset':
                 return 'boolean'
-        return JsdocsParser.getFunctionReturnType(self, name)
+        return JsdocsParser.getFunctionReturnType(self, name, retval)
+
+
+class JsdocsCPP(JsdocsParser):
+    def setupSettings(self):
+        nameToken = '[a-zA-Z_][a-zA-Z0-9_]*'
+        identifier = '(%s)(::%s)?' % (nameToken, nameToken)
+        self.settings = {
+            'typeInfo': False,
+            'curlyTypes': False,
+            'typeTag': 'param',
+            'commentCloser': ' */',
+            'fnIdentifier': identifier,
+            'varIdentifier': identifier,
+            'fnOpener': identifier + '\\s+' + identifier + '\\s*\\(',
+            'bool': 'bool',
+            'function': 'function'
+        }
+
+    def parseFunction(self, line):
+        res = re.search(
+            '(?P<retval>' + self.settings['varIdentifier'] + ')[&*\\s]+'
+            + '(?P<name>' + self.settings['varIdentifier'] + ')'
+            # void fnName
+            # (arg1, arg2)
+            + '\\s*\\(\\s*(?P<args>.*)\)',
+            line
+        )
+        if not res:
+            return None
+
+        return (res.group('name'), res.group('args'), res.group('retval'))
+
+    def parseArgs(self, args):
+        if args.strip() == 'void':
+            return []
+        return super(JsdocsCPP, self).parseArgs(args)
+
+    def getArgType(self, arg):
+        return None
+
+    def getArgName(self, arg):
+        return re.search("(" + self.settings['varIdentifier'] + r")(?:\s*\[\s*\])?(?:\s*=.*)?$", arg).group(1)
+
+    def parseVar(self, line):
+        return None
+
+    def guessTypeFromValue(self, val):
+        return None
+
+    def getFunctionReturnType(self, name, retval):
+        return retval if retval != 'void' else None
 
 
 class JsdocsCoffee(JsdocsParser):
     def setupSettings(self):
+        identifier = '[a-zA-Z_$][a-zA-Z_$0-9]*'
         self.settings = {
             # curly brackets around the type information
             'curlyTypes': True,
             'typeTag': "type",
             'typeInfo': True,
             # technically, they can contain all sorts of unicode, but w/e
-            'varIdentifier': '[a-zA-Z_$][a-zA-Z_$0-9]*',
-            'fnIdentifier': '[a-zA-Z_$][a-zA-Z_$0-9]*',
+            'varIdentifier': identifier,
+            'fnIdentifier': identifier,
+            'fnOpener': None,  # no multi-line function definitions for you, hipsters!
             'commentCloser': '###',
             'bool': 'Boolean',
             'function': 'Function'
@@ -520,10 +648,10 @@ class JsdocsCoffee(JsdocsParser):
             return None
 
         # grab the name out of "name1 = function name2(foo)" preferring name1
-        name = escape(res.group('name') or '')
+        name = res.group('name') or ''
         args = res.group('args')
 
-        return (name, args)
+        return (name, args, None)
 
     def parseVar(self, line):
         res = re.search(
@@ -572,6 +700,7 @@ class JsdocsActionscript(JsdocsParser):
             'commentCloser': ' */',
             'fnIdentifier': nameToken,
             'varIdentifier': '(%s)(?::%s)?' % (nameToken, nameToken),
+            'fnOpener': 'function(?:\\s+[gs]et)?(?:\\s+' + nameToken + ')?\\s*\\(',
             'bool': 'bool',
             'function': 'function'
         }
@@ -584,7 +713,7 @@ class JsdocsActionscript(JsdocsParser):
             # function fnName
             + '(?:\s+(?P<name2>' + self.settings['fnIdentifier'] + '))?'
             # (arg1, arg2)
-            + '\s*\((?P<args>.*)\)',
+            + '\s*\(\s*(?P<args>.*)\)',
             line
         )
         if not res:
@@ -599,7 +728,7 @@ class JsdocsActionscript(JsdocsParser):
         if res.group('getset') == 'set':
             options['as_setter'] = True
 
-        return (name, args, options)
+        return (name, args, None, options)
 
     def parseVar(self, line):
         return None
@@ -609,6 +738,93 @@ class JsdocsActionscript(JsdocsParser):
 
     def getArgType(self, arg):
         # could actually figure it out easily, but it's not important for the documentation
+        return None
+
+
+class JsdocsObjC(JsdocsParser):
+
+    def setupSettings(self):
+        identifier = '[a-zA-Z_$][a-zA-Z_$0-9]*'
+        self.settings = {
+            # curly brackets around the type information
+            "curlyTypes": True,
+            'typeInfo': True,
+            "typeTag": "type",
+            # technically, they can contain all sorts of unicode, but w/e
+            "varIdentifier": identifier,
+            "fnIdentifier":  identifier,
+            "fnOpener": '^\s*[-+]',
+            "commentCloser": " */",
+            "bool": "Boolean",
+            "function": "Function"
+        }
+
+    def getDefinition(self, view, pos):
+        maxLines = 25  # don't go further than this
+
+        definition = ''
+        for i in xrange(0, maxLines):
+            line = read_line(view, pos)
+            if line is None:
+                break
+
+            pos += len(line) + 1
+            # strip comments
+            line = re.sub("//.*", "", line)
+            if definition == '':
+                if not self.settings['fnOpener'] or not re.search(self.settings['fnOpener'], line):
+                    definition = line
+                    break
+            definition += line
+            if line.find(';') > -1 or line.find('{') > -1:
+                definition = re.sub(r'\s*[;{]\s*$', '', definition)
+                break
+        return definition
+
+    def parseFunction(self, line):
+        # this is terrible, don't judge me
+
+        typeRE = r'[a-zA-Z_$][a-zA-Z0-9_$]*\s*\**'
+        res = re.search(
+            '[-+]\s+\\(\\s*(?P<retval>' + typeRE + ')\\s*\\)\\s*'
+            + '(?P<name>[a-zA-Z_$][a-zA-Z0-9_$]*)'
+            # void fnName
+            # (arg1, arg2)
+            + '\\s*(?::(?P<args>.*))?',
+            line
+        )
+        if not res:
+            return
+        name = res.group('name')
+        argStr = res.group('args')
+        args = []
+        if argStr:
+            groups = re.split('\\s*:\\s*', argStr)
+            numGroups = len(groups)
+            for i in xrange(0, numGroups):
+                group = groups[i]
+                if i < numGroups - 1:
+                    result = re.search(r'\s+(\S*)$', group)
+                    name += ':' + result.group(1)
+                    group = group[:result.start()]
+
+                args.append(group)
+
+            if (numGroups):
+                name += ':'
+        return (name, '|||'.join(args), res.group('retval'))
+
+    def parseArgs(self, args):
+        out = []
+        for arg in args.split('|||'):  # lol
+            lastParen = arg.rfind(')')
+            out.append((arg[1:lastParen], arg[lastParen + 1:]))
+        return out
+
+    def getFunctionReturnType(self, name, retval):
+        return retval if retval != 'void' and retval != 'IBAction' else None
+
+    def parseVar(self, line):
         return None
 
 
@@ -639,8 +855,10 @@ class JsdocsIndentCommand(sublime_plugin.TextCommand):
             v.insert(edit, currPos, "\t")
 
     def getIndentSpaces(self, line):
-        res = re.search("^\\s*\\*(?P<fromStar>\\s*@(?:param|property)\\s+\\S+\\s+\\S+\\s+)\\S", line) \
-           or re.search("^\\s*\\*(?P<fromStar>\\s*@(?:returns?|define)\\s+\\S+\\s+)\\S", line) \
+        hasTypes = getParser(self.view).settings['typeInfo']
+        extraIndent = '\\s+\\S+' if hasTypes else ''
+        res = re.search("^\\s*\\*(?P<fromStar>\\s*@(?:param|property)%s\\s+\\S+\\s+)\\S" % extraIndent, line) \
+           or re.search("^\\s*\\*(?P<fromStar>\\s*@(?:returns?|define)%s\\s+\\S+\\s+)\\S" % extraIndent, line) \
            or re.search("^\\s*\\*(?P<fromStar>\\s*@[a-z]+\\s+)\\S", line) \
            or re.search("^\\s*\\*(?P<fromStar>\\s*)", line)
         if res:
@@ -731,3 +949,110 @@ class JsdocsTrimAutoWhitespace(sublime_plugin.TextCommand):
         line = v.substr(lineRegion)
         spaces = max(0, v.settings().get("jsdocs_indentation_spaces", 1))
         v.replace(edit, lineRegion, re.sub("^(\\s*\\*)\\s*$", "\\1\n\\1" + (" " * spaces), line))
+
+
+class JsdocsWrapLines(sublime_plugin.TextCommand):
+    """
+    Reformat description text inside a comment block to wrap at the correct length.
+    Wrap column is set by the first ruler (set in Default.sublime-settings), or 80 by default.
+    Shortcut Key: alt+q
+    """
+
+    def run(self, edit):
+        v = self.view
+        settings = v.settings()
+        rulers = settings.get('rulers')
+        tabSize = settings.get('tab_size')
+
+        wrapLength = rulers[0] or 80
+        numIndentSpaces = max(0, settings.get("jsdocs_indentation_spaces", 1))
+        indentSpaces = " " * numIndentSpaces
+        indentSpacesSamePara = " " * max(0, settings.get("jsdocs_indentation_spaces_same_para", numIndentSpaces))
+        spacerBetweenSections = settings.get("jsdocs_spacer_between_sections")
+
+        v.run_command('expand_selection', {'to': 'scope'})
+
+
+        # find the first word
+        startPoint = v.find("\n\\s*\\* ", v.sel()[0].begin()).begin()
+        # find the first tag, or the end of the comment
+        endPoint = v.find("\\s*\n\\s*\\*(/)", v.sel()[0].begin()).begin()
+
+        # replace the selection with this ^ new selection
+        v.sel().clear()
+        v.sel().add(sublime.Region(startPoint, endPoint))
+
+        # get the description text
+        text = v.substr(v.sel()[0])
+
+        # find the indentation level
+        indentation = len(re.sub('\t', ' ' * tabSize, re.search("\n(\\s*\\*)", text).group(1)))
+        wrapLength -= indentation - tabSize
+
+        # join all the lines, collapsing "empty" lines
+        text = re.sub("\n(\\s*\\*\\s*\n)+", "\n\n", text)
+
+        def wrapPara(para):
+            para = re.sub("(\n|^)\\s*\\*\\s*", " ", para)
+
+            # split the paragraph into words
+            words = para.strip().split(' ')
+            text = '\n'
+            line = ' *' + indentSpaces
+            lineTagged = False # indicates if the line contains a doc tag
+            paraTagged = False # indicates if this paragraph contains a doc tag
+            lineIsNew = True
+            tag = ''
+
+            # join all words to create lines, no longer than wrapLength
+            for i, word in enumerate(words):
+                if not word and not lineTagged:
+                    continue
+
+                if lineIsNew and word[0] == '@':
+                    lineTagged = True
+                    paraTagged = True
+                    tag = word
+
+                if len(line) + len(word) >= wrapLength - 1:
+                    # appending the word to the current line whould exceed its
+                    # length requirements
+                    text += line.rstrip() + '\n'
+                    line = ' *' + indentSpacesSamePara + word + ' '
+                    lineTagged = False
+                    lineIsNew = True
+                else:
+                    line += word + ' '
+
+                lineIsNew = False
+
+            text += line.rstrip()
+            return {'text':text,
+                    'lineTagged':lineTagged,
+                    'tagged':paraTagged,
+                    'tag':tag}
+
+        # split the text into paragraphs, where each paragraph is eighter
+        # defined by an empty line or the start of a doc parameter
+        paragraphs = re.split('\n{2,}|\n\\s*\\*\\s*(?=@)', text)
+        wrappedParas = []
+        text = ''
+        for p, para in enumerate(paragraphs):
+            # wrap the lines in the current paragraph
+            wrappedParas.append(wrapPara(para))
+
+        # combine all the paragraphs into a single piece of text
+        for i in range(0, len(wrappedParas)):
+            para = wrappedParas[i]
+            last = i == len(wrappedParas) - 1
+
+            nextIsTagged = not last and wrappedParas[i+1]['tagged']
+            nextIsSameTag = nextIsTagged and para['tag'] == wrappedParas[i+1]['tag']
+
+            if last or (para['lineTagged'] or nextIsTagged) and \
+                    not (spacerBetweenSections and not nextIsSameTag):
+                text += para['text']
+            else:
+                text += para['text'] + '\n *'
+
+        write(v, text)